Extend Quill component
