+++ cms/modules/src/properties/xhtml/xhtml.property.ts
@@ -1,9 +1,9 @@
 import { Component, ViewChild } from '@angular/core';
-import Quill from 'quill'
-import { QuillEditorComponent } from 'ngx-quill';
-
 import { CmsProperty, PAGE_TYPE, MEDIA_TYPE } from '@angular-cms/core';
 import { DropEvent } from '../../shared/drag-drop/drop-event.model';
+import { QuillEditorComponent } from 'ngx-quill';
+
+import './ngx-quill.extension';
 
 @Component({
     selector: '[xhtmlProperty]',
@@ -22,20 +22,12 @@ import { DropEvent } from '../../shared/drag-drop/drop-event.model';
     `
 })
 export class XHtmlProperty extends CmsProperty {
-    @ViewChild('editor', {
-        static: true
-    }) editor: QuillEditorComponent
+    @ViewChild('editor', { static: true }) editor: QuillEditorComponent;
 
-    constructor() {
-        super();
-        //TODO how using inline style in quill editor
-        console.log(Quill.imports);
-    }
 
     onDropItem(e: DropEvent) {
-        const { type, extendProperties } = e.dragData;
-        const typeOfContent = extendProperties ? extendProperties.type : type;
-        switch (typeOfContent) {
+        const { type } = e.dragData;
+        switch (type) {
             case PAGE_TYPE:
                 this.insertPageUrl(e.dragData);
                 break;
@@ -51,25 +43,14 @@ export class XHtmlProperty extends CmsProperty {
         const insertOps = [
             { insert: { image: src }, attributes: { width: '150' } }
         ]
-        this.quillInsertAtCursorPosition(this.editor.quillEditor, insertOps);
+        this.editor.insertAtCursorPosition(insertOps);
     }
 
     private insertPageUrl(dragData) {
-        const { name, extendProperties } = dragData;
+        const { name, linkUrl } = dragData;
         const insertOps = [
-            { insert: name, attributes: { link: extendProperties.linkUrl } }
+            { insert: name, attributes: { link: linkUrl } }
         ]
-        this.quillInsertAtCursorPosition(this.editor.quillEditor, insertOps);
-    }
-
-    private quillInsertAtCursorPosition(quillEditor: any, insertOperators: { insert: any, attributes: any }[]) {
-        if (!quillEditor) return;
-
-        const operators: any[] = [];
-        const range = quillEditor.getSelection(true);
-        if (range.index > 0) operators.push({ retain: range.index });
-
-        insertOperators.forEach(insert => operators.push(insert));
-        quillEditor.updateContents({ ops: operators });
+        this.editor.insertAtCursorPosition(insertOps);
     }
 }
